[WIP] Adding distance class
This is meant to respond to Issue #13.
As mention in the issue, this basically just uses code from
https://github.com/ryanmrichard/ForceManII/blob/master/ForceManII/InternalCoords/Distance.cpp

This compiles, but the code needs to be rearranged a bit, and something should
happen with Util.hpp and LibChemistDefs.hpp--either they are combined, or
added to a different part of the code, or something.

Also still need a test case and Doxygen.